remove hytek from route
closes #11
