+++ src/routes/meets/[meetId]/+layout.js
@@ -1,29 +1,35 @@
+import { meetsMap } from '$lib/services/meetsService.js';
 import { error } from '@sveltejs/kit';
 
-/**
- * Load meet data and make it available to child routes
- *
- * @param fetch
- * @param params
- * @param parent
- * @returns {Promise<{meetConfig: *}>}
- */
-export async function load({ fetch, params }) {
-  const { meetId } = params;
+export async function load({ fetch, params, parent }) {
+  const data = await parent();
+  const { meetId, eventId } = params;
+  const { meetConfig } = data ;
 
+  // get the event list
   try {
-    const res = await fetch(`/meets/${meetId}/config`);
-    const { meetConfig } = await res.json();
+    // early check if meet exists in DB
+    if (!meetsMap.has(meetId)) {
+      throw error(404, `Event with meetId '${meetId}' not found.`);
+    }
 
-    if (res.ok) {
-      return {
-        // pass down the meet config to child layouts
-        meetConfig
-      };
-    } else {
-      throw error(404);
+    const getEvtIndex = await fetch(`/meets/${meetId}/evtindex`);
+    const { evtIndexHTML } = await getEvtIndex.json();
+
+    return {
+      meetConfig: meetsMap.get(meetId),
+      evtIndexHTML,
+      meetId,
+      eventId,
     }
   } catch (e) {
-    console.error('Error fetching meet config', e);
+    console.warn('Could not fetch event index', e);
+    // todo: should warn user there is a config problem (event list will never load)
+    return {
+      meetConfig,
+      meetId,
+      eventId,
+      evtIndexHTML: null
+    }
   }
 }
