+++ src/routes/meets/[meetId]/event/[eventId]/+error.svelte
