+++ src/routes/meets/[meetId]/event/[eventId]/+page.js
@@ -6,7 +6,7 @@ export async function load({ fetch, params }) {
   try {
     // call page endpoint manually since there is already a load function in layout-hytek
     // do not include .html extension since this causes Vercel to render the wrong html
-    const evtRequest = await fetch(`/meets/${meetId}/hytek/event/${eventId}-html`);
+    const evtRequest = await fetch(`/meets/${meetId}/event/${eventId}-html`);
 
     if (!evtRequest.ok) {
       if (evtRequest.status === 404) {
