+++ src/routes/meets/[meetId]/hytek/+server.js
@@ -1,0 +1,6 @@
+import { redirect } from '@sveltejs/kit';
+
+export function GET({ url }) {
+  console.log({ url })
+  throw redirect(307, url.pathname.replace(/\/hytek$/, ''));
+}
